test(mcp): CheckErrorsProbe.cs is deleted, its last 28 facts now carried by CheckErrorsFamilyTests (#486)
Phase 2 close. Every fact CheckErrorsProbe.cs asserted through the deleted 1.x
single-family renderer now has a fresh test in CheckErrorsFamilyTests.cs against
the merged Wire.RenderCheck / JsonWire.RenderCheck, so the probe file carries no
assertion this branch still owes.

Before deleting it: ScriptPropertyCheckProbe.cs called three of its json-parity
helpers (JsonMatches, JsonNull, JsonHasHistogram) across nine call sites. Deleting
the file out from under them would have added fresh compile errors to a file
phase 3 converts, not phase 2 — so the three helpers move to
ScriptPropertyCheckProbe.cs itself first (it already fails to build for its own,
unrelated, reason: the same phase-1 (b) sites owed to phase 3). Verified: the
solution's error count drops by exactly CheckErrorsProbe.cs's 110 lines (186 -> 76
sites... rebuilt: CheckMergeProbe 2, EpochGuardProbe 16, DialogueInfoOrderProbe 6,
BulkPrimitivesWave3Probe 14, DialogueValidateGuardProbe 3, ScriptPropertyCheckProbe
28 — the last one unchanged from before this commit, confirming the relocation
added no new errors there.

harness-residue-baseline.json drops the CheckErrorsProbe.cs key (150 -> 149 probe
files). CheckErrorsFamilyTests.cs already carries the Converted-from marker
HarnessResidueTests' one-way-conversion guard requires.

Part of #486.
