+++ src/housecarl-mcp-tests/harness-residue-baseline.json
@@ -27,7 +27,6 @@
     "src/housecarl-generator/BulkPrimitivesWave2Probe.cs": { "lines": 225, "ciAllRows": 1 },
     "src/housecarl-generator/BulkPrimitivesWave3Probe.cs": { "lines": 891, "ciAllRows": 1 },
     "src/housecarl-generator/BulkQueryPrimitivesProbe.cs": { "lines": 404, "ciAllRows": 1 },
-    "src/housecarl-generator/CheckErrorsProbe.cs": { "lines": 2436, "ciAllRows": 1 },
     "src/housecarl-generator/CheckMergeProbe.cs": { "lines": 2014, "ciAllRows": 1 },
     "src/housecarl-generator/ClosureCopyProbe.cs": { "lines": 226, "ciAllRows": 1 },
     "src/housecarl-generator/ClosureWalkProbe.cs": { "lines": 294, "ciAllRows": 1 },
